requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
absl-py==2.1.0
    # via tensorboard
accelerate==1.2.1
    # via
    #   sentence-transformers
    #   transformers
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   openai
asttokens==3.0.0
    # via stack-data
attrs==24.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
datasets==3.2.0
    # via sentence-transformers
debugpy==1.8.11
    # via ipykernel
decorator==5.1.1
    # via ipython
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
distro==1.9.0
    # via openai
executing==2.1.0
    # via stack-data
filelock==3.16.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
grpcio==1.68.1
    # via tensorboard
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
huggingface-hub==0.27.0
    # via
    #   accelerate
    #   datasets
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipykernel==6.29.5
    # via -r requirements.in
ipython==8.31.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.1.5
    # via -r requirements.in
jedi==0.19.2
    # via ipython
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
    #   torch
jiter==0.8.2
    # via openai
joblib==1.4.2
    # via scikit-learn
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
jupyterlab-widgets==3.0.13
    # via ipywidgets
markdown==3.7
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
narwhals==1.19.1
    # via altair
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.4.2
    # via torch
numpy==2.2.1
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   tensorboard
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
openai==1.58.1
    # via -r requirements.in
packaging==24.2
    # via
    #   accelerate
    #   altair
    #   datasets
    #   huggingface-hub
    #   ipykernel
    #   streamlit
    #   tensorboard
    #   transformers
pandas==2.2.3
    # via
    #   -r requirements.in
    #   datasets
    #   streamlit
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.0.0
    # via
    #   sentence-transformers
    #   streamlit
platformdirs==4.3.6
    # via jupyter-core
prompt-toolkit==3.0.48
    # via ipython
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.2
    # via
    #   streamlit
    #   tensorboard
psutil==6.1.1
    # via
    #   accelerate
    #   ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==18.1.0
    # via
    #   datasets
    #   streamlit
pydantic==2.10.4
    # via openai
pydantic-core==2.27.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via
    #   ipython
    #   rich
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
pyzmq==26.2.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   streamlit
    #   transformers
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
safetensors==0.4.5
    # via
    #   accelerate
    #   transformers
scikit-learn==1.6.0
    # via sentence-transformers
scipy==1.14.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.3.1
    # via -r requirements.in
setuptools==75.6.0
    # via
    #   tensorboard
    #   torch
six==1.17.0
    # via
    #   python-dateutil
    #   tensorboard
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
stack-data==0.6.3
    # via ipython
streamlit==1.41.1
    # via -r requirements.in
sympy==1.13.1
    # via torch
tenacity==9.0.0
    # via streamlit
tensorboard==2.18.0
    # via -r requirements.in
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.5.1
    # via
    #   -r requirements.in
    #   accelerate
    #   sentence-transformers
    #   transformers
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
transformers==4.47.1
    # via
    #   -r requirements.in
    #   sentence-transformers
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via tensorboard
widgetsnbextension==4.0.13
    # via ipywidgets
xxhash==3.5.0
    # via datasets
yarl==1.18.3
    # via aiohttp
